+++ ft_printf/utils0.c
@@ -60,10 +60,10 @@ int	ft_print_addr(void *p0)
 
 int	check_unsigned(va_list args)
 {
-	unsigned long long	n;
-	int					count;
+	unsigned int	n;
+	int				count;
 
-	n = va_arg(args, unsigned long long);
+	n = va_arg(args, unsigned int);
 	count = count_unsigned_digits(n);
 	ft_putnbr_fd_unsigned(n, 1);
 	return (count);
